+++ src/graphql/resolver/create_resolver_function/get_operation_type/mod.rs
@@ -1,0 +1,26 @@
+use log::debug;
+
+use crate::{
+    configuration::subgraph::entities::service_entity_field::ServiceEntityField,
+    graphql::{resolver::ServiceResolver, schema::ResolverType},
+};
+
+impl ServiceResolver {
+    pub fn get_operation_type(
+        resolver_type: &ResolverType,
+        as_field: &Option<ServiceEntityField>,
+    ) -> ResolverType {
+        debug!("Getting Operation Type For Resolver Type");
+        match as_field {
+            Some(as_field) => {
+                if as_field.list.unwrap_or(false) {
+                    ResolverType::FindMany
+                } else {
+                    ResolverType::FindOne
+                }
+            }
+            None => resolver_type.clone(),
+        }
+    }
+}
+
